+++ tests/js/arrow1.test.ts
@@ -1,8 +1,8 @@
-import * as test from "tape";
 import * as wasm from "../../pkg/node/arrow1";
 import { readFileSync } from "fs";
 import { tableFromIPC, tableToIPC } from "apache-arrow";
 import { testArrowTablesEqual, readExpectedArrowData, temporaryServer } from "./utils";
+import { describe, it, expect } from "vitest";
 
 // Path from repo root
 const dataDir = "tests/data";
@@ -21,70 +21,64 @@ const testFiles = [
   "2-partition-zstd.parquet",
 ];
 
-test("read file", async (t) => {
+describe("read file", async (t) => {
   const expectedTable = readExpectedArrowData();
 
   for (const testFile of testFiles) {
-    const dataPath = `${dataDir}/${testFile}`;
-    const arr = new Uint8Array(readFileSync(dataPath));
-    const table = tableFromIPC(wasm.readParquet(arr).intoIPC());
-    testArrowTablesEqual(t, expectedTable, table);
+    it(testFile, () => {
+      const dataPath = `${dataDir}/${testFile}`;
+      const arr = new Uint8Array(readFileSync(dataPath));
+      const table = tableFromIPC(wasm.readParquet(arr).intoIPCStream());
+      testArrowTablesEqual(expectedTable, table);
+    });
   }
-
-  t.end();
 });
 
-test("read-write-read round trip (with writer properties)", async (t) => {
+it("read-write-read round trip (with writer properties)", async (t) => {
   const dataPath = `${dataDir}/1-partition-brotli.parquet`;
   const buffer = readFileSync(dataPath);
   const arr = new Uint8Array(buffer);
-  const initialTable = tableFromIPC(wasm.readParquet(arr).intoIPC());
+  const initialTable = tableFromIPC(wasm.readParquet(arr).intoIPCStream());
 
   const writerProperties = new wasm.WriterPropertiesBuilder().build();
 
   const parquetBuffer = wasm.writeParquet(
-    wasm.Table.fromIPC(tableToIPC(initialTable, "stream")),
+    wasm.Table.fromIPCStream(tableToIPC(initialTable, "stream")),
     writerProperties
   );
-  const table = tableFromIPC(wasm.readParquet(parquetBuffer).intoIPC());
+  const table = tableFromIPC(wasm.readParquet(parquetBuffer).intoIPCStream());
 
-  testArrowTablesEqual(t, initialTable, table);
-  t.end();
+  testArrowTablesEqual(initialTable, table);
 });
 
-test("read-write-read round trip (no writer properties provided)", async (t) => {
+it("read-write-read round trip (no writer properties provided)", async (t) => {
   const dataPath = `${dataDir}/1-partition-brotli.parquet`;
   const buffer = readFileSync(dataPath);
   const arr = new Uint8Array(buffer);
-  const initialTable = tableFromIPC(wasm.readParquet(arr).intoIPC());
+  const initialTable = tableFromIPC(wasm.readParquet(arr).intoIPCStream());
 
   const parquetBuffer = wasm.writeParquet(
-    wasm.Table.fromIPC(tableToIPC(initialTable, "stream"))
+    wasm.Table.fromIPCStream(tableToIPC(initialTable, "stream"))
   );
-  const table = tableFromIPC(wasm.readParquet(parquetBuffer).intoIPC());
+  const table = tableFromIPC(wasm.readParquet(parquetBuffer).intoIPCStream());
 
-  testArrowTablesEqual(t, initialTable, table);
-  t.end();
+  testArrowTablesEqual(initialTable, table);
 });
 
-test("error produced trying to read file with arrayBuffer", (t) => {
+it("error produced trying to read file with arrayBuffer", (t) => {
   const arrayBuffer = new ArrayBuffer(10);
   try {
     // @ts-expect-error input should be Uint8Array
     wasm.readParquet(arrayBuffer);
   } catch (err) {
-    t.ok(err instanceof Error, "err expected to be an Error");
-    t.equals(
-      err.message,
-      "Empty input provided or not a Uint8Array.",
-      "Expected error message"
+    expect(err instanceof Error, "err expected to be an Error").toBeTruthy();
+    expect(err.message, "Expected error message").toStrictEqual(
+      "Empty input provided or not a Uint8Array."
     );
   }
-
-  t.end();
 });
 
-test("read stream-write stream-read stream round trip (no writer properties provided)", async (t) => {
+it("read stream-write stream-read stream round trip (no writer properties provided)", async (t) => {
   const server = await temporaryServer();
   const listeningPort = server.addresses()[0].port;
   const rootUrl = `http://localhost:${listeningPort}`;
@@ -96,9 +90,8 @@ test("read stream-write stream-read stream round trip (no writer properties prov
 
   const stream = await wasm.transformParquetStream(originalStream);
   const accumulatedBuffer = new Uint8Array(await new Response(stream).arrayBuffer());
-  const roundtripTable = tableFromIPC(wasm.readParquet(accumulatedBuffer).intoIPC());
+  const roundtripTable = tableFromIPC(wasm.readParquet(accumulatedBuffer).intoIPCStream());
 
-  testArrowTablesEqual(t, expectedTable, roundtripTable);
+  testArrowTablesEqual(expectedTable, roundtripTable);
   await server.close();
-  t.end();
 })