+++ tests/js/index.test.ts
@@ -1,0 +1,5 @@
+import "./arrow1.test";
+import "./arrow1-ffi.test";
+import "./arrow2.test";
+import "./arrow2-geo-metadata.test";
+import "./arrow2-ffi.test";